[refactor]: 불필요한 훅을 유틸 함수로 변경 (#73)
